# packages/client/spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Ownware Gateway — core wire contract
  version: 0.29.0
  description: >
    Drive a Ownware agent from any language: start a run, stream its
    events, answer permission prompts. Bearer auth applies to every
    route except /health whenever the gateway has auth enabled (always
    true on non-loopback binds; token at <dataDir>/gateway-token).
servers:
  - url: http://localhost:3011
    description: Default local gateway (ownware serve prints yours)
security:
  - bearer: []
paths:
...
  /api/v1/sources:
    post:
      operationId: registerSource
      summary: Register one logical source using safe control metadata only.
      description: >
        Requires a delegated principal; workspace and profile scope come only
        from the verified bearer. The request cannot carry paths, URLs, bytes,
        storage locators, or workspace/profile authority. Equal metadata under
        distinct idempotency keys remains distinct logical sources.
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceRegistrationInput' }
      responses:
        '202':
          description: Pending safe manifest. No content or storage detail is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceManifest' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/SourceCreationConflictResponse' }
        '413': { $ref: '#/components/responses/ErrorResponse' }
        '500': { $ref: '#/components/responses/ErrorResponse' }
    get:
      operationId: listSources
      summary: List one bounded page of safe manifests in the verified scope.
      description: Requires a delegated principal. An empty scope returns an empty page.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: cursor
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bounded source page without raw content or storage detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceList' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/sources/{sourceId}:
    get:
      operationId: getSource
      summary: Read one safe source manifest in the verified scope.
      description: Cross-workspace/profile identities are indistinguishable from absence.
      parameters:
        - in: path
          name: sourceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Safe source control manifest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceManifest' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/sources/{sourceId}/versions/{sourceVersionId}/jobs:
    post:
      operationId: createSourceJob
      summary: Enqueue one allowlisted operation for an exact immutable source version.
      description: >
        Workspace and profile scope come only from the delegated bearer. Creation
        requires a UUID idempotency key; an exact completed retry replays the
        original safe job projection across Gateway restarts.
      parameters:
        - { in: path, name: sourceId, required: true, schema: { type: string, format: uuid } }
        - { in: path, name: sourceVersionId, required: true, schema: { type: string, format: uuid } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operation]
              properties:
                operation: { type: string, const: inspect_format }
              additionalProperties: false
      responses:
        '202':
          description: Durable job accepted. The response contains no worker or storage detail.
          content: { application/json: { schema: { $ref: '#/components/schemas/SourceJob' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/SourceCreationConflictResponse' }
...
  /api/v1/source-jobs/{jobId}:
    get:
      operationId: getSourceJob
      summary: Read one scoped safe source-job projection.
      description: Cross-workspace/profile identities are indistinguishable from absence.
      parameters:
        - { in: path, name: jobId, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Current durable state without claims, leases, retry times, paths, content, or raw diagnostics.
          content: { application/json: { schema: { $ref: '#/components/schemas/SourceJob' } } }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  responses:
    ErrorResponse:
      description: Safe typed error. Raw internal/provider detail is not returned.
      headers:
        X-Ownware-Correlation-Id:
          schema: { type: string, format: uuid }
          description: Opaque server-generated support correlation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    SourceCreationConflictResponse:
      description: >
        Safe typed creation conflict. `source_quota_exceeded` reports only the
        blocked resource class plus standard safe error metadata; it never
        reports usage, failed workspace/profile scope, resource identities,
        paths, or a retry delay.
      headers:
        X-Ownware-Correlation-Id:
          schema: { type: string, format: uuid }
          description: Opaque server-generated support correlation.
      content:
        application/json:
          schema:
            oneOf:
              - allOf:
                  - { $ref: '#/components/schemas/ErrorResponse' }
                  - not:
                      required: [error]
                      properties:
                        error: { const: source_quota_exceeded }
              - { $ref: '#/components/schemas/SourceQuotaExceededError' }
...
  schemas:
...
    SourceRegistrationInput:
      type: object
      required: [kind, label, classification, authority, audiencePolicyRef, sensitivityPolicyRef, purposePolicyRef, retentionPolicyRef, freshnessPolicyRef]
      properties:
        kind:
          type: string
          enum: [file, text, visual, structured_export, cloud_document, connected_snapshot, supported_other]
        label:
          type: string
          minLength: 1
          maxLength: 160
          description: Safe single-line display label; control characters reject.
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example, excluded] }
        audiencePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        sensitivityPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        purposePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        retentionPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshnessPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
      additionalProperties: false
    SourceHealth:
      type: object
      required: [registration, inspection, preparation, access, freshness, conflict, deletion]
      properties:
        registration: { type: string, enum: [pending, registered, rejected] }
        inspection: { type: string, enum: [not_started, queued, inspecting, complete, partial, failed] }
        preparation: { type: string, enum: [not_requested, queued, preparing, ready, partial, failed] }
        access: { type: string, enum: [available, denied, expired, disconnected, wrong_identity] }
        freshness: { type: string, enum: [fresh, aging, stale, unknown] }
        conflict: { type: string, enum: [none, suspected, confirmed, resolved] }
        deletion: { type: string, enum: [active, frozen, deleting, partially_deleted, deleted] }
      additionalProperties: false
    SourceManifest:
      type: object
      required: [sourceId, kind, label, classification, authority, audiencePolicyRef, sensitivityPolicyRef, purposePolicyRef, retentionPolicyRef, freshnessPolicyRef, revision, currentVersionId, health, createdAt, updatedAt]
      properties:
        sourceId: { type: string, format: uuid }
        kind:
          type: string
          enum: [file, text, visual, structured_export, cloud_document, connected_snapshot, supported_other]
        label: { type: string, minLength: 1, maxLength: 160 }
        classification: { type: string, enum: [public, internal, confidential, restricted] }
        authority: { type: string, enum: [source_of_record, supporting_reference, example, excluded] }
        audiencePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        sensitivityPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        purposePolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        retentionPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        freshnessPolicyRef: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
        revision: { type: integer, minimum: 1 }
        currentVersionId: { type: [string, 'null'], format: uuid }
        health: { $ref: '#/components/schemas/SourceHealth' }
        createdAt: { type: integer, minimum: 0 }
        updatedAt: { type: integer, minimum: 0 }
      additionalProperties: false
    SourceList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/SourceManifest' }
        nextCursor: { type: [string, 'null'], format: uuid }
      additionalProperties: false
...
    SourceJob:
      type: object
      required: [jobId, sourceId, sourceVersionId, operation, implementationVersion, resourceId, dataViewId, state, attempt, maxAttempts, checkpoint, cancelRequestedAt, outcomeCode, createdAt, updatedAt, terminalAt]
      properties:
        jobId: { type: string, format: uuid }
        sourceId: { type: string, format: uuid }
        sourceVersionId: { type: string, format: uuid }
        operation: { type: string, enum: [inspect_format, extract_text, prepare_data_view] }
        implementationVersion: { type: string, enum: [inspect_format.v1, text_extraction.v1, csv_data_view.v1] }
        resourceId:
          type: [string, 'null']
          format: uuid
          description: Published only after successful preparation creates its safe resource manifest.
        dataViewId:
          type: [string, 'null']
          format: uuid
          description: Published only after successful Data View preparation; no cells or private locator are returned.
        state:
          type: string
          enum: [queued, running, waiting_for_resource, cancel_requested, succeeded, partial, failed, cancelled]
        attempt: { type: integer, minimum: 0, maximum: 3 }
        maxAttempts: { type: integer, const: 3 }
        checkpoint: { type: integer, minimum: 0, maximum: 4 }
        cancelRequestedAt: { type: [integer, 'null'], minimum: 0 }
        outcomeCode:
          type: [string, 'null']
          enum: [attempts_exhausted, cancelled, inspection_complete, inspection_timeout, inspection_unavailable, source_format_invalid, source_object_mismatch, source_object_missing, source_object_oversized, source_storage_inconsistent, preparation_complete, preparation_timeout, preparation_unavailable, data_view_unavailable, data_view_invalid, data_view_too_large, data_view_publication_conflict, artifact_cleanup_failed, csv_identity_invalid, csv_input_oversized, csv_invalid_utf8, csv_header_empty, csv_header_duplicate, csv_quoting_invalid, csv_bare_carriage_return, csv_row_ragged, csv_field_limit_exceeded, csv_row_limit_exceeded, csv_cell_limit_exceeded, csv_total_cell_limit_exceeded, csv_preparation_timeout, null]
        createdAt: { type: integer, minimum: 0 }
        updatedAt: { type: integer, minimum: 0 }
        terminalAt: { type: [integer, 'null'], minimum: 0 }
        cancellation: { type: string, enum: [requested, already_requested] }
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, message, category, correlationId]
      properties:
        error: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        message: { type: string, maxLength: 500 }
        category: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        correlationId: { type: string, format: uuid }
        retryAfter:
          type: number
          minimum: 0
          description: Optional retry delay in seconds.
        actualRevision:
          type: integer
          minimum: 1
          description: Safe actual source revision, present only for `source_upload_refresh_conflict`.
        actualCurrentVersionId:
          type: [string, 'null']
          format: uuid
          description: Safe actual current source version, present only for `source_upload_refresh_conflict`.
...
    SourceQuotaExceededError:
      type: object
      required: [error, message, category, correlationId, resourceClass]
      description: >
        Quota denial without usage, failed scope, source/upload/job/resource
        identities, paths, or retry timing.
      properties:
        error: { type: string, const: source_quota_exceeded }
        message: { type: string, const: Source quota does not allow this operation. }
        category: { type: string, const: invalid_request }
        correlationId: { type: string, format: uuid }
        resourceClass:
          type: string
          enum: [source_registrations, source_storage_bytes, source_upload_sessions, source_jobs, source_derived_resources]
      additionalProperties: false
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Gateway session token. Enabled automatically on any non-loopback
        bind; persisted at <dataDir>/gateway-token (0600).